Remove redundant casts

+++ src/main/java/ldbc/snb/datagen/dictionary/FlashmobTagDictionary.java
@@ -136,7 +136,7 @@ public class FlashmobTagDictionary {
         double currentProb = 0.0;
         for (int i = 0; i < size; ++i) {
             flashmobTagCumDist[i].prob = currentProb;
-            currentProb += (double) (flashmobTagCumDist[i].level) / (double) (sumLevels);
+            currentProb += (double) (flashmobTagCumDist[i].level) / sumLevels;
         }
     }
